requirements.txt
pillow>=10.0.0
python-dotenv>=1.0.1
ffmpeg-python>=0.2.0
# UI
customtkinter>=5.2.2

# From your imports
opencv-python>=4.10.0
numpy>=1.26
tqdm>=4.66

# Nude detection
nudenet>=3.0.0

ImageHash>=4.3.1
Pillow>=10.3
onnxruntime>=1.18,<2
onnxruntime-gpu>=1.18,<2


